requirements.txt
openai==1.12.0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.1.9
langchain-openai==0.0.8
chromadb==0.4.22 